Cleanup and improvements
Package remux goes away in favour of urlparser. No need to have regex
based mux anymore since there are many other and better implementations
out there.

The urlparser is a simple mapper of variable arguments in the URL to a
string map.

Old unusable templates go away as well. Copyrights are updated and a new
AUTHORS file is instroduced.